Rename file to webPush.js
